+++ lucene/core/src/resources/META-INF/services/org.apache.lucene.codecs.KnnVectorsFormat
@@ -16,4 +16,3 @@
 org.apache.lucene.codecs.lucene99.Lucene99HnswVectorsFormat
 org.apache.lucene.codecs.lucene104.Lucene104ScalarQuantizedVectorsFormat
 org.apache.lucene.codecs.lucene104.Lucene104HnswScalarQuantizedVectorsFormat
-org.apache.lucene.codecs.lucene106.dedup.Lucene106DedupHnswVectorsFormat
