+++ lucene/sandbox/src/java/org/apache/lucene/sandbox/codecs/dedup/DedupHnswVectorsFormat.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.codecs.lucene106.dedup;
+package org.apache.lucene.sandbox.codecs.dedup;
 
 import static org.apache.lucene.codecs.lucene99.Lucene99HnswVectorsFormat.DEFAULT_BEAM_WIDTH;
 import static org.apache.lucene.codecs.lucene99.Lucene99HnswVectorsFormat.DEFAULT_MAX_CONN;
@@ -61,8 +61,8 @@ import org.apache.lucene.util.hnsw.HnswGraph;
  *
  * @lucene.experimental
  */
-public final class Lucene106DedupHnswVectorsFormat extends KnnVectorsFormat {
-  private static final String NAME = "Lucene106DedupHnswVectorsFormat";
+public final class DedupHnswVectorsFormat extends KnnVectorsFormat {
+  private static final String NAME = "DedupHnswVectorsFormat";
 
   /**
    * Controls how many of the nearest neighbor candidates are connected to the new node. Defaults to
@@ -99,7 +99,7 @@ public final class Lucene106DedupHnswVectorsFormat extends KnnVectorsFormat {
   private final int tinySegmentsThreshold;
 
   /** Constructs a format using default graph construction parameters */
-  public Lucene106DedupHnswVectorsFormat() {
+  public DedupHnswVectorsFormat() {
     this(
         DEFAULT_MAX_CONN, DEFAULT_BEAM_WIDTH, DEFAULT_NUM_MERGE_WORKER, null, HNSW_GRAPH_THRESHOLD);
   }
@@ -110,7 +110,7 @@ public final class Lucene106DedupHnswVectorsFormat extends KnnVectorsFormat {
    * @param maxConn the maximum number of connections to a node in the HNSW graph
    * @param beamWidth the size of the queue maintained during graph construction.
    */
-  public Lucene106DedupHnswVectorsFormat(int maxConn, int beamWidth) {
+  public DedupHnswVectorsFormat(int maxConn, int beamWidth) {
     this(maxConn, beamWidth, DEFAULT_NUM_MERGE_WORKER, null, HNSW_GRAPH_THRESHOLD);
   }
 
@@ -122,7 +122,7 @@ public final class Lucene106DedupHnswVectorsFormat extends KnnVectorsFormat {
    * @param tinySegmentsThreshold the expected number of vector operations to return k nearest
    *     neighbors of the current graph size
    */
-  public Lucene106DedupHnswVectorsFormat(int maxConn, int beamWidth, int tinySegmentsThreshold) {
+  public DedupHnswVectorsFormat(int maxConn, int beamWidth, int tinySegmentsThreshold) {
     this(maxConn, beamWidth, DEFAULT_NUM_MERGE_WORKER, null, tinySegmentsThreshold);
   }
 
@@ -137,7 +137,7 @@ public final class Lucene106DedupHnswVectorsFormat extends KnnVectorsFormat {
    *     generated by this format to do the merge. If null, the configured {@link
    *     MergeScheduler#getIntraMergeExecutor(MergePolicy.OneMerge)} is used.
    */
-  public Lucene106DedupHnswVectorsFormat(
+  public DedupHnswVectorsFormat(
       int maxConn, int beamWidth, int numMergeWorkers, ExecutorService mergeExec) {
     this(maxConn, beamWidth, numMergeWorkers, mergeExec, HNSW_GRAPH_THRESHOLD);
   }
@@ -155,7 +155,7 @@ public final class Lucene106DedupHnswVectorsFormat extends KnnVectorsFormat {
    * @param tinySegmentsThreshold the expected number of vector operations to return k nearest
    *     neighbors of the current graph size
    */
-  public Lucene106DedupHnswVectorsFormat(
+  public DedupHnswVectorsFormat(
       int maxConn,
       int beamWidth,
       int numMergeWorkers,
