+++ lucene/sandbox/src/resources/META-INF/services/org.apache.lucene.codecs.KnnVectorsFormat
@@ -14,3 +14,4 @@
 #  limitations under the License.
 
 org.apache.lucene.sandbox.codecs.faiss.FaissKnnVectorsFormat
+org.apache.lucene.sandbox.codecs.dedup.DedupHnswVectorsFormat
